+++ core-ng/src/test/java/core/framework/internal/http/RetryInterceptorTest.java
@@ -48,9 +48,105 @@ class RetryInterceptorTest {
         assertThat(interceptor.shouldRetry(1, "POST", new HttpConnectTimeoutException("connection timeout"))).isTrue();
         assertThat(interceptor.shouldRetry(1, "POST", new ConnectException("connection refused"))).isTrue();
 
+        /* connect timeout stack trace
+        Caused by: java.net.SocketTimeoutException: connect timed out
+            at java.base/java.net.PlainSocketImpl.socketConnect(Native Method)
+            at java.base/java.net.AbstractPlainSocketImpl.doConnect(Unknown Source)
+            at java.base/java.net.AbstractPlainSocketImpl.connectToAddress(Unknown Source)
+            at java.base/java.net.AbstractPlainSocketImpl.connect(Unknown Source)
+            at java.base/java.net.SocksSocketImpl.connect(Unknown Source)
+            at java.base/java.net.Socket.connect(Unknown Source)
+            at okhttp3.internal.platform.Platform.connectSocket(Platform.kt:127)
+            at okhttp3.internal.connection.RealConnection.connectSocket(RealConnection.kt:268)
+            at okhttp3.internal.connection.RealConnection.connect(RealConnection.kt:176)
+            at okhttp3.internal.connection.ExchangeFinder.findConnection(ExchangeFinder.kt:236)
+            at okhttp3.internal.connection.ExchangeFinder.findHealthyConnection(ExchangeFinder.kt:109)
+        */
+        assertThat(interceptor.shouldRetry(1, "POST", new SocketTimeoutException("connect timed out"))).isTrue();
+    }
+
+    /* Read timeout stack trace with http 1.1
+    Caused by: java.net.SocketTimeoutException: timeout
+        at okio.SocketAsyncTimeout.newTimeoutException(Okio.kt:159)
+        at okio.AsyncTimeout.exit$jvm(AsyncTimeout.kt:203)
+        at okio.AsyncTimeout$source$1.read(AsyncTimeout.kt:163)
+        at okio.RealBufferedSource.indexOf(RealBufferedSource.kt:349)
+        at okio.RealBufferedSource.readUtf8LineStrict(RealBufferedSource.kt:222)
+        at okhttp3.internal.http1.Http1ExchangeCodec.readHeaderLine(Http1ExchangeCodec.kt:210)
+        at okhttp3.internal.http1.Http1ExchangeCodec.readResponseHeaders(Http1ExchangeCodec.kt:181)
+        at okhttp3.internal.connection.Exchange.readResponseHeaders(Exchange.kt:105)
+        at okhttp3.internal.http.CallServerInterceptor.intercept(CallServerInterceptor.kt:82)
+        at okhttp3.internal.http.RealInterceptorChain.proceed(RealInterceptorChain.kt:112)
+        at okhttp3.internal.http.RealInterceptorChain.proceed(RealInterceptorChain.kt:87)
+        at core.framework.internal.http.ServiceUnavailableInterceptor.intercept(ServiceUnavailableInterceptor.java:19)
+        at okhttp3.internal.http.RealInterceptorChain.proceed(RealInterceptorChain.kt:112)
+        at okhttp3.internal.connection.ConnectInterceptor.intercept(ConnectInterceptor.kt:37)
+        at okhttp3.internal.http.RealInterceptorChain.proceed(RealInterceptorChain.kt:112)
+        at okhttp3.internal.http.RealInterceptorChain.proceed(RealInterceptorChain.kt:87)
+        at okhttp3.internal.cache.CacheInterceptor.intercept(CacheInterceptor.kt:82)
+        at okhttp3.internal.http.RealInterceptorChain.proceed(RealInterceptorChain.kt:112)
+        at okhttp3.internal.http.RealInterceptorChain.proceed(RealInterceptorChain.kt:87)
+        at okhttp3.internal.http.BridgeInterceptor.intercept(BridgeInterceptor.kt:84)
+        at okhttp3.internal.http.RealInterceptorChain.proceed(RealInterceptorChain.kt:112)
+        at okhttp3.internal.http.RetryAndFollowUpInterceptor.intercept(RetryAndFollowUpInterceptor.kt:71)
+        at okhttp3.internal.http.RealInterceptorChain.proceed(RealInterceptorChain.kt:112)
+        at okhttp3.internal.http.RealInterceptorChain.proceed(RealInterceptorChain.kt:87)
+        at core.framework.internal.http.RetryInterceptor.intercept(RetryInterceptor.java:38)
+        at okhttp3.internal.http.RealInterceptorChain.proceed(RealInterceptorChain.kt:112)
+        at okhttp3.internal.http.RealInterceptorChain.proceed(RealInterceptorChain.kt:87)
+        at okhttp3.RealCall.getResponseWithInterceptorChain(RealCall.kt:184)
+        at okhttp3.RealCall.execute(RealCall.kt:66)
+        at core.framework.internal.http.HTTPClientImpl.execute(HTTPClientImpl.java:49)
+        ... 2 more
+    Caused by: java.net.SocketTimeoutException: Read timed out
+        at java.base/java.net.SocketInputStream.socketRead0(Native Method)
+        at java.base/java.net.SocketInputStream.socketRead(SocketInputStream.java:115)
+        at java.base/java.net.SocketInputStream.read(SocketInputStream.java:168)
+        at java.base/java.net.SocketInputStream.read(SocketInputStream.java:140)
+        at okio.InputStreamSource.read(Okio.kt:102)
+        at okio.AsyncTimeout$source$1.read(AsyncTimeout.kt:159)
+        ... 29 more
+     */
+    /* Read timeout stack trace with http 2.0
+    Caused by: java.net.SocketTimeoutException: timeout
+        at okhttp3.internal.http2.Http2Stream$StreamTimeout.newTimeoutException(Http2Stream.kt:666)
+        at okhttp3.internal.http2.Http2Stream$StreamTimeout.exitAndThrowIfTimedOut(Http2Stream.kt:675)
+        at okhttp3.internal.http2.Http2Stream.takeHeaders(Http2Stream.kt:142)
+        at okhttp3.internal.http2.Http2ExchangeCodec.readResponseHeaders(Http2ExchangeCodec.kt:99)
+        at okhttp3.internal.connection.Exchange.readResponseHeaders(Exchange.kt:105)
+        at okhttp3.internal.http.CallServerInterceptor.intercept(CallServerInterceptor.kt:82)
+        at okhttp3.internal.http.RealInterceptorChain.proceed(RealInterceptorChain.kt:112)
+        at okhttp3.internal.http.RealInterceptorChain.proceed(RealInterceptorChain.kt:87)
+        at core.framework.internal.http.ServiceUnavailableInterceptor.intercept(ServiceUnavailableInterceptor.java:19)
+        at okhttp3.internal.http.RealInterceptorChain.proceed(RealInterceptorChain.kt:112)
+        at okhttp3.internal.connection.ConnectInterceptor.intercept(ConnectInterceptor.kt:37)
+        at okhttp3.internal.http.RealInterceptorChain.proceed(RealInterceptorChain.kt:112)
+        at okhttp3.internal.http.RealInterceptorChain.proceed(RealInterceptorChain.kt:87)
+        at okhttp3.internal.cache.CacheInterceptor.intercept(CacheInterceptor.kt:82)
+        at okhttp3.internal.http.RealInterceptorChain.proceed(RealInterceptorChain.kt:112)
+        at okhttp3.internal.http.RealInterceptorChain.proceed(RealInterceptorChain.kt:87)
+        at okhttp3.internal.http.BridgeInterceptor.intercept(BridgeInterceptor.kt:84)
+        at okhttp3.internal.http.RealInterceptorChain.proceed(RealInterceptorChain.kt:112)
+        at okhttp3.internal.http.RetryAndFollowUpInterceptor.intercept(RetryAndFollowUpInterceptor.kt:71)
+        at okhttp3.internal.http.RealInterceptorChain.proceed(RealInterceptorChain.kt:112)
+        at okhttp3.internal.http.RealInterceptorChain.proceed(RealInterceptorChain.kt:87)
+        at core.framework.internal.http.RetryInterceptor.intercept(RetryInterceptor.java:38)
+        at okhttp3.internal.http.RealInterceptorChain.proceed(RealInterceptorChain.kt:112)
+        at okhttp3.internal.http.RealInterceptorChain.proceed(RealInterceptorChain.kt:87)
+        at okhttp3.RealCall.getResponseWithInterceptorChain(RealCall.kt:184)
+        at okhttp3.RealCall.execute(RealCall.kt:66)
+        at core.framework.internal.http.HTTPClientImpl.execute(HTTPClientImpl.java:49)
+        ... 2 more
+     */
+    @Test
+    void shouldRetryWithReadTimeout() {
         // socket read timeout caught by AsyncTimeout
         var timeout = new SocketTimeoutException("timeout");
         timeout.initCause(new SocketTimeoutException("Read timed out"));
+        assertThat(interceptor.shouldRetry(1, "POST", timeout)).isFalse();
+        assertThat(interceptor.shouldRetry(2, "PUT", timeout)).isTrue();
+
+        timeout = new SocketTimeoutException("timeout");
         assertThat(interceptor.shouldRetry(1, "POST", timeout)).isFalse();
         assertThat(interceptor.shouldRetry(2, "PUT", timeout)).isTrue();
 
@@ -61,7 +157,7 @@ class RetryInterceptorTest {
         assertThat(interceptor.shouldRetry(2, "PUT", timeout)).isTrue();
 
         // okio AsyncTimout cancels call when if call timout
-        IOException cancelled = new IOException("Canceled");
+        var cancelled = new IOException("Canceled");
         assertThat(interceptor.shouldRetry(1, "POST", cancelled)).isFalse();
         assertThat(interceptor.shouldRetry(2, "PUT", cancelled)).isTrue();
     }